chore(jws): rename Jwt into Jws
