+++ tests/Unit/Jws/GeneratorConfigTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Imdhemy\AppStore\Tests\Unit\Jwt;
+namespace Imdhemy\AppStore\Tests\Unit\Jws;
 
 use DateTimeImmutable;
-use Imdhemy\AppStore\Jwt\GeneratorConfig;
-use Imdhemy\AppStore\Jwt\Issuer;
-use Imdhemy\AppStore\Jwt\Key;
+use Imdhemy\AppStore\Jws\GeneratorConfig;
+use Imdhemy\AppStore\Jws\Issuer;
+use Imdhemy\AppStore\Jws\Key;
 use Imdhemy\AppStore\Tests\TestCase;
 use Lcobucci\Clock\FrozenClock;
 use Lcobucci\JWT\Signer\Ecdsa\Sha256;
