+++ tests/Unit/Jws/IssuerTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Imdhemy\AppStore\Tests\Unit\Jwt;
+namespace Imdhemy\AppStore\Tests\Unit\Jws;
 
-use Imdhemy\AppStore\Jwt\Issuer;
-use Imdhemy\AppStore\Jwt\Key;
+use Imdhemy\AppStore\Jws\Issuer;
+use Imdhemy\AppStore\Jws\Key;
 use Imdhemy\AppStore\Tests\TestCase;
 use Lcobucci\JWT\Signer\Key\InMemory;
 use Lcobucci\JWT\Signer\Rsa\Sha256;
